project/requirements.txt
altgraph==0.17.4
bcrypt==4.3.0
blinker==1.9.0
brotlicffi==1.0.9.2
certifi==2025.8.3
cffi==1.15.0
charset-normalizer==3.3.2
click==8.2.1
cryptography==45.0.7
et-xmlfile==2.0.0
Flask==3.1.1
greenlet==3.2.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.1
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
paramiko==2.11.0
PyMySQL==1.1.2
PyNaCl==1.6.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0
requests==2.32.4
SQLAlchemy==2.0.43
sshtunnel==0.4.0
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==4.0.2
Werkzeug==3.1.3
pypandoc